Add inputs, outputs, evaluate, variableRegex
