+++ src/js/variableRegex.mjs
@@ -1,0 +1,2 @@
+// Matches the first character of a variable
+export default /[A-Z]/i;
